+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Version latest (2019-08-07)
 
 - ENH: moved path of aws catalog from gitRemoteAws.pull into dotman.fn.awsCat
+- BUGFIX: refactor `df` to `df_pd` to distnguish from `df_json` in `pull`
+- BUGFIX: flattening the `pricing` field from `www.ec2instance.info` was yielding NaN. Replaced with the average price of all regions.
 
 
 Version 0.3.5 (2019-08-07)